ParamManager and new quantization framework (#477)
This PR introduces parameter manager, which keeps information of all
parameters throughout the entire model, and is in charge of applying
quantization and dequantization to the model. Now all parameters are
managed in a unified and deterministic way, and the quantization
computation will no longer be affected by the optimization passes.

Together with the ParamManager, this PR does an overhaul of the
quantization framework with more modular and customizable quantization
modules. Specifically, we have

- QuantizationScheme, which describes how an entire model is quantized,
and
- QuantizationSpec, which describes how a particular parameter is
quantized.

With them and the ParamManager, we now decoupled the quantization
algorithm description with the model transformation, which provides
us with better customizability for quantization.

We will further support loading models whose weights are pre-quantized.
This PR sets up the framework, while some followup changes are needed
to fully support this functionality.

Please refer to `mlc_llm/relax_model/param_manager.py` for more
documentation on the design and implementation of the ParamManager,
and to `mlc_llm/quantization/quantization.py` for more documentation
on the new quantization framework.

-----

This PR also adds two fusion passes which are essential to performance.
They are required because the previous quantization framework operates
on subgraph level, while the new quantization framework only works on
single relax.Var level. The previous quantization framework directly
rewrites some subgraphs into fused pattern, which is not realizable in
the new framework, making the fusion passes needed.

-----

To enable other ongoing works that were blocked by ParamManager, this PR
**only supports LLaMA with ParamManager and the latest quantization
framework**. We will update other model architectures accordingly soon.
Meanwhile, this PR marks that we starts a stage where both the old and
new quantization framework co-exist. We will clean the old one once all
model architectures are updated with ParamManager and the new
quantization framework.

This PR keeps the compilation flow of other model architectures
unaffected. Performance of RedPajama and RWKV are tested locally through
CLI, and their performance are not affected by this PR.