[Fix] Lib with small `max_seq_len` incompatible with prebuilt weight (#840)
This PR fixes an issue introduced by #780, which broke our intended
behavior to make the cos/sin shape independent of the max sequence
length, so that no matter what max sequence length people use, they
can always use a same set of prebuilt weight and do not need to
clone different weight repositories. This intended behavior is broken
by #780.

However, it is true that the needs for larger max sequence length are
growing. Prior to #780, when the max sequence length is larger than
2048, the cached cos/sin do not work anymore and break. To be compatible
as much as possible, this PR changes the behavior to "taking the
maximum value of 2048 and the specified max sequence length when
building the model lib".

With this fix, when the maximum sequence length is smaller than 2048,
we are still able to use the prebuilt weights. And when it is larger
than 2048, we will only be able to use the weight converted along the
build.