[Fix] Fix ft quantization in param manager (#917)
This PR fixes the ft quantization error in param manager.

Previously in `ParamManager.get_quantized_param_info`, it calls 
```python
f_quantize = param.quant_spec.get_quantize_func(param_info)
```
And in `FTRowwiseQuantizationSpec.get_quantize_func`, it calls `bb.emit_te` which requires a relax function frame in the `BlockBuilder` stack. But in the `ParamManager.get_quantized_param_info` it does not have a scope of `with bb.function`. So I add this scope to fix this bug.