Add Mixtral and GPT-BigCode to Model Preset (#1621)
In SLM stack, we have a convention in `mlc_chat compile` that it comes
with builtin model architectures, so that compilation could happen
without depending on actually cloning an HF repo. We have been doing
this for most of the models we support, including:
- GPT2
- Mistral
- Llama
- Phi

However, I noticed just now that GPT-BigCode and Mixtral are missing,
thanks to @Hzfengsy's reminders. This PR adds them to the model preset.