[Model] Remove redundant `batch_forward` and move broadcast (#1900)
This PR contains four changes:

1. It removes the duplicate `batch_forward` defined in model
definitions. This function was widely used prior to our migration
to PagedKVCache, since before migration the attention codepath
of single sequence forward and batch forward differ. But since our
migration, the codepaths are unified into one, and therefore we
can safely remove most `batch_forward` functions.

2. It moves `op.ccl_broadcast_from_worker0` from model main forward
(which will be called at the beginning of prefill/decode) to embedding.
This change has two benefits. Firstly, the token ids taken by `embed`
was not broadcasted across workers, and it is possible for workers
other than 0 to have illegal token ids which is not in the range of
vocab size, and moving the broadcasting to `embed` perfectly address
this issue. Secondly, broadcasting token ids in `embed` is more
lightweight than broadcasting embeddings in `prefill`/`decode`, since
the tensor size of token ids is much smaller.

3. It adds `max_batch_size` to the config class of models, so that
they are potentially compatible with batching and MLC serve.

4. It removes the `k_cache` and `v_cache` effects from the models
that have switched to PagedKVCache support.

Randomly picked a few models (as below) to run the engine test, and
all of them are passed:

* phi-2 with tp=2,
* RedPajama with tp=2,
* stablelm with tp=2 (since stablelm does not support TP right now).