Deprecate old flow (#1928)
* Deprecate old flow

This PR deprecates the old flow.
As of today most of the efforts are centralized around the new flow
with SLM compilation. Additionally, we are bringing model definitions
through unified kv interface so we can have a single model
across all backends, server and local setting.

We kept the old flow around for a while, but it is a good
time to do the transition. All the documents are updated
to point to the new flow.

We also created a backup branch
https://github.com/mlc-ai/mlc-llm/tree/backup-before-old-flow-deprecation
for people who would like to checkout some of the old flow references.

* Remove deprecated prebuilts