[KVCache] Support matching arbitrary element offset for aux data (#2057)
This PR enhances the TIR attention-related functions to support
matching arbitrary element offests. This makes room for the KV cache
to allocate a large array the all the auxiliary data and do slicing
on it.

This PR should affect nothing for the current codebase, given all
the element offsets are zeros as of now.