[Serving][Refactor] Python interface refactor (#2085)
This PR is an initial major Python interface refactor of MLC Serve.

With this PR, `mlc_llm.serve` in Python now exposes two engine classes:
`AsyncEngine` and `Engine`. Both classes have two entrypoints,
`chat_completion` and `completion` which conform to OpenAI Python API
(reference: https://github.com/openai/openai-python).

As the name suggested, `AsyncEngine` works asynchronously, and `Engine`
works synchronously. It worths noting that the `Engine` since this PR
is different from the `Engine` so far. The new `Engine` does not provide
interfaces for batch generation.

For robustness and correctness, the old `Engine` in Python is moved
to `mlc_llm.serve.sync_engine.SyncEngine`. We do not directly expose
this SyncEngine, and it now mainly serves testing and debug purposes.
It is useful to check the correctness of new features, because of its
simplicity. It keeps the low-level interface to directly invoke `step()`
function of the engine, and also keeps the low-level batch generation
interface.

Our REST API entry points defined under `mlc_llm/serve/entrypoints/`
are also refactored accordingly to adapt to the latest Python API
in MLC Serve. In short, most of the logic in OpenAI API entry points
are moved to Python API, which simplifies the implementation of
entry points.

Please note that this is the first (also the largest) planned refactor.
We will follow up with some other refactors, which have smaller scopes
compared with this PR. The planned refactors include:

* provide submodule interface to align OpenAI Python package in
https://github.com/openai/openai-python
* refactor the constructor interface of `Engine`/`AsyncEngine` to
align the MLC serve CLI interface.