[Serving][Spec] Fix the output inconsistent bug of q0f32 spec decoding (#2184)
- According to https://github.com/mlc-ai/mlc-llm/issues/2167, the problem that
the output of spec decoding in q0f32 is inconsistent with the single
model of q0f32 has been fixed. 
- Modified the test_engine_generate function located in
`tests/python/serve/test_serve_engine_spec.py` to support comparison
of the output of a single model and the output of spec decoding
- The accuracy comparison with hugging face is left (because
the current version of llama-2-7b of q0f32 cannot be consistent
with the output of hugging face model)
- The output of spec decoding for q0f16 cannot be consistent
with the output of a single model of q0f16, but this may be due
to floating point errors.   

Co-authored-by: DearFishi <yw6m20@soton.ac.u>