[Sampler] Fix GPU sampler behavior when batch size is 0 (#2234)
This PR adds the early exit for the GPU sampler, which ran into
GPU kernels even when the batch size is 0 prior to this commit.

The 0 batch size case can happen when parallel generation of a request
and engine preemption exists. In this case, the GPU sampler should
just synchronization and return, and not run into any GPU kernel.