[Pass] Attach manual softmax-with-temperature (#2280)
This PR updates all the models to use the new softmax-with-temperature
function, which inlines the temperature division (or argmax if
temperature is 0) process into the two-stage softmax.

Unit benchmark shows that the inline of division does no harm to the
softmax. When batch size is large, the inlined softmax can have better
performance than a standalone divide kernel, which takes much time
when batch size is large.