[iOS] Introducing package CLI for iOS app packaging (#2297)
This PR introduces the packaging CLI `mlc_llm package` which
reads from a `mlc-package-config.json` and compiles model
and prepares model/runtime libraries automatically.

With this PR, we get rid of prebuilt model library dependency
for iOS app build.

Validated that the iOS build can work. iOS documentation is updated
according to this latest change. The same flow is supposed to work
for Android as well, while it still needs verification for Android
app build.