[Sampler] Enhance checks for whether FlashInfer is enabled (#2502)
This PR improves the check in GPU sampler for whether FlashInfer is
enabled. Previously we did not check the CUDA compute capability,
which makes the GPU sampler not able to properly run on Colab where
the T4 GPU has a compute version of 7.5 which FlashInfer does not
support.

With this PR, when the compute capability is less than 8.0, we
will not use FlashInfer in GPU sampler.