[PrefixCache] Defer sequence extension (#2654)
This PR deferrs the prefix cache sequence extention.
Previously, the prefix cache update is committed after every action,
which is unnecessary. We can defer this sequence extention and
commit the extentions when the prefix cache is used again.

This PR also changes the IntTuple used in PrefixCache to
`std::vector<int32_t>` for less data structure construction overhead.